# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexus POS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  category: string;
  imageUrl?: string | null;
  barcode?: string | null;
  inStock: boolean;
  stockCount: number;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  description?: string;
  price: number;
  category: string;
  imageUrl?: string;
  barcode?: string;
  inStock?: boolean;
  stockCount?: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
  refunded: "refunded",
  voided: "voided",
} as const;

export type OrderDiscountType =
  | (typeof OrderDiscountType)[keyof typeof OrderDiscountType]
  | null;

export const OrderDiscountType = {
  percent: "percent",
  fixed: "fixed",
} as const;

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  discountAmount?: number | null;
  lineTotal: number;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  subtotal: number;
  discountType?: OrderDiscountType;
  discountAmount?: number | null;
  discountValue?: number | null;
  tax: number;
  total: number;
  paymentMethod?: string | null;
  splitCardAmount?: number | null;
  splitCashAmount?: number | null;
  notes?: string | null;
  voidReason?: string | null;
  customerId?: number | null;
  items: OrderItem[];
  createdAt: string;
  completedAt?: string | null;
}

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
  discountAmount?: number;
};

export type CreateOrderBodyDiscountType =
  (typeof CreateOrderBodyDiscountType)[keyof typeof CreateOrderBodyDiscountType];

export const CreateOrderBodyDiscountType = {
  percent: "percent",
  fixed: "fixed",
} as const;

export interface CreateOrderBody {
  items: CreateOrderBodyItemsItem[];
  paymentMethod: string;
  splitCardAmount?: number;
  splitCashAmount?: number;
  discountType?: CreateOrderBodyDiscountType;
  discountAmount?: number;
  notes?: string;
  customerId?: number;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
  refunded: "refunded",
  voided: "voided",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
  voidReason?: string;
}

export type HeldOrderDiscountType =
  | (typeof HeldOrderDiscountType)[keyof typeof HeldOrderDiscountType]
  | null;

export const HeldOrderDiscountType = {
  percent: "percent",
  fixed: "fixed",
} as const;

export interface HeldOrderItem {
  productId: number;
  productName: string;
  price: number;
  quantity: number;
}

export interface HeldOrder {
  id: number;
  label?: string | null;
  items: HeldOrderItem[];
  notes?: string | null;
  discountType?: HeldOrderDiscountType;
  discountAmount?: number | null;
  createdAt: string;
}

export type CreateHeldOrderBodyDiscountType =
  (typeof CreateHeldOrderBodyDiscountType)[keyof typeof CreateHeldOrderBodyDiscountType];

export const CreateHeldOrderBodyDiscountType = {
  percent: "percent",
  fixed: "fixed",
} as const;

export interface CreateHeldOrderBody {
  label?: string;
  items: HeldOrderItem[];
  notes?: string;
  discountType?: CreateHeldOrderBodyDiscountType;
  discountAmount?: number;
}

export interface DashboardSummary {
  todaySales: number;
  todayOrders: number;
  totalProducts: number;
  avgOrderValue: number;
  weekSales: number;
  weekOrders: number;
}

export interface CategorySales {
  category: string;
  totalSales: number;
  orderCount: number;
}

export interface DailySales {
  date: string;
  revenue: number;
  orders: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalRevenue: number;
  unitsSold: number;
}

export interface PaymentMethodSales {
  method: string;
  revenue: number;
  count: number;
}

export interface Customer {
  id: number;
  name: string;
  email?: string | null;
  phone?: string | null;
  loyaltyPoints: number;
  totalSpent: number;
  orderCount: number;
  createdAt: string;
}

export interface CreateCustomerBody {
  name: string;
  email?: string;
  phone?: string;
}

export interface ReportSummary {
  revenue: number;
  orders: number;
  avgOrderValue: number;
  newCustomers: number;
  topProduct?: string | null;
  voidedOrders: number;
}

export interface HourlySales {
  hour: number;
  revenue: number;
  orders: number;
}

export interface ListCustomersQueryParams {
  search?: string;
}

export interface GetReportSummaryQueryParams {
  from?: string;
  to?: string;
}

export interface GetHourlySalesQueryParams {
  date?: string;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  limit?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
  refunded: "refunded",
  voided: "voided",
} as const;

export type GetRecentOrdersParams = {
  limit?: number;
};

export type GetDailySalesParams = {
  days?: number;
};

export type GetTopProductsParams = {
  limit?: number;
};

export type ListCustomersParams = {
  search?: string;
};

export type GetLowStockProductsParams = {
  threshold?: number;
};

export type GetReportSummaryParams = {
  from?: string;
  to?: string;
};

export type GetHourlySalesParams = {
  date?: string;
};

export type ExportOrdersParams = {
  from?: string;
  to?: string;
};