# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [dashboard]
      summary: Get recent orders for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /dashboard/daily-sales:
    get:
      operationId: getDailySales
      tags: [dashboard]
      summary: Get daily revenue for the last N days
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daily sales data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailySales"
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top selling products by revenue
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/orders:
    get:
      operationId: getCustomerOrders
      tags: [customers]
      summary: Get all orders for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /dashboard/low-stock:
    get:
      operationId: getLowStockProducts
      tags: [dashboard]
      summary: Get products with low stock
      parameters:
        - name: threshold
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Low stock products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        category:
          type: string
        imageUrl:
          type: string
          nullable: true
        barcode:
          type: string
          nullable: true
        inStock:
          type: boolean
        stockCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - category
        - inStock
        - stockCount
        - createdAt
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled, refunded, voided]
        subtotal:
          type: number
          format: float
        discountType:
          type: string
          enum: [percent, fixed]
          nullable: true
        discountAmount:
          type: number
          format: float
          nullable: true
        discountValue:
          type: number
          format: float
          nullable: true
        tax:
          type: number
          format: float
        total:
          type: number
          format: float
        paymentMethod:
          type: string
          nullable: true
        splitCardAmount:
          type: number
          format: float
          nullable: true
        splitCashAmount:
          type: number
          format: float
          nullable: true
        notes:
          type: string
          nullable: true
        voidReason:
          type: string
          nullable: true
        customerId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - orderNumber
        - status
        - subtotal
        - tax
        - total
        - items
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        discountAmount:
          type: number
          format: float
          nullable: true
        lineTotal:
          type: number
          format: float
      required:
        - id
        - productId
        - productName
        - quantity
        - unitPrice
        - lineTotal
...
    DailySales:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
          format: float
        orders:
          type: integer
      required:
        - date
        - revenue
        - orders
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalRevenue:
          type: number
          format: float
        unitsSold:
          type: integer
      required:
        - productId
        - productName
        - totalRevenue
        - unitsSold
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        loyaltyPoints:
          type: integer
        totalSpent:
          type: number
          format: float
        orderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - loyaltyPoints
        - totalSpent
        - orderCount
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
      required:
        - name
...