# lib/api-zod/src/generated/types/createCustomerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexus POS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCustomerBody {
  name: string;
  email?: string;
  phone?: string;
}