# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexus POS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  email?: string | null;
  phone?: string | null;
  loyaltyPoints: number;
  totalSpent: number;
  orderCount: number;
  createdAt: Date;
}